feat: Add IP geolocation to transactions
Detect and store customer's approximate location (city, region) during checkout using IP geolocation. This data is captured non-intrusively and added to the transaction details in the database.

The customer's location will now be displayed in the transaction receipt and in the admin customer manager. The `UserDetails` type has been updated to include `customerLocation`.